Fix memory issue
1) reimplementing the Esp Methods
2) removing any `std::string` to avoid the memory issue
3) adding a function to convert form the integer to a string `Helper_IntToString`
4) fixing a bug in the Usart.cpp in the boad-rate method
5) Adding a test file to test the Esp

+++ dev/mcal/src/Usart.cpp
@@ -67,7 +67,7 @@ void Usart::_SetBaudRate() {
     const uint32_t scale = 16;
 
     // Calculate the USARTDIV value
-    float UsartDiv = ((clockFrequency) / (scale * config_.baudRate));
+    float UsartDiv = (static_cast<float>(clockFrequency) / (scale * config_.baudRate));
 
     // Extract the mantissa and fraction parts
     uint16_t divMantissa  = (uint16_t)UsartDiv;
